requirements.txt
PySide6>=6.8,<7.0
pywin32>=308
Pillow>=11.1.0
onnxruntime>=1.20.0
numpy>=2.2.0
pyinstaller>=6.12.0
pytest>=8.3.0
